+++ src/main/java/ru/kata/spring/boot_security/demo/model/Role.java
@@ -10,15 +10,15 @@ public class Role implements GrantedAuthority {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
     private String name;
 
-    public Role(int id, String name) {
+    public Role(long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Role(int id) {
+    public Role(long id) {
         this.id = id;
     }
 
@@ -29,11 +29,11 @@ public class Role implements GrantedAuthority {
     public Role() {
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
